requirements.txt
openai
whisper
fastapi
uvicorn
pydub
soundfile
python-dotenv
aiofiles
streamlit